make unsupported sim item look like a valid sim but with different content
