+++ app/src/main/java/com/hover/stax/presentation/sim/components/SimItem.kt
@@ -1,0 +1,207 @@
+package com.hover.stax.presentation.sim.components
+
+import android.content.Context
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import coil.compose.AsyncImage
+import coil.request.CachePolicy
+import coil.request.ImageRequest
+import com.hover.stax.R
+import com.hover.stax.domain.model.Account
+import com.hover.stax.presentation.home.BalanceTapListener
+import com.hover.stax.presentation.home.TopBar
+import com.hover.stax.ui.theme.*
+import com.hover.stax.utils.DateUtils
+
+private fun isSimSupportedAccount(accountId: Int) : Boolean = accountId >=0
+
+@Composable
+internal fun SimItem(
+	simIndex: Int,
+	account: Account,
+	bonus: Int,
+	secondaryClickItem: () -> Unit,
+	balanceTapListener: BalanceTapListener?
+) {
+	val size13 = dimensionResource(id = R.dimen.margin_13)
+
+	Box(
+		modifier = Modifier
+			.fillMaxWidth()
+			.padding(vertical = size13)
+			.shadow(elevation = 0.dp)
+			.border(width = 1.dp, color = DarkGray, shape = RoundedCornerShape(5.dp))
+	) {
+		Column(
+			modifier = Modifier
+				.fillMaxWidth()
+				.padding(size13)
+		) {
+			SimItemTopRow(
+				simIndex = simIndex,
+				account = account,
+				balanceTapListener = balanceTapListener
+			)
+			Column {
+				val notYetChecked = stringResource(id = R.string.not_yet_checked)
+				Text(
+					text = stringResource(
+						id = R.string.airtime_balance_holder,
+						account.latestBalance ?: notYetChecked
+					),
+					color = TextGrey,
+					modifier = Modifier.padding(top = size13),
+					style = MaterialTheme.typography.body1
+				)
+
+				Text(
+					text = stringResource(
+						id = R.string.as_of,
+						DateUtils.humanFriendlyDateTime(account.latestBalanceTimestamp)
+					),
+					color = TextGrey,
+					modifier = Modifier.padding(bottom = 26.dp),
+					style = MaterialTheme.typography.body1
+				)
+
+				OutlinedButton(
+					onClick = secondaryClickItem,
+					modifier = Modifier
+						.padding(bottom = 6.dp)
+						.shadow(elevation = 0.dp)
+						.wrapContentWidth(),
+					shape = MaterialTheme.shapes.medium,
+					border = BorderStroke(width = 0.5.dp, color = DarkGray),
+					colors = ButtonDefaults.buttonColors(
+						backgroundColor = ColorSurface,
+						contentColor = OffWhite
+					)
+				) {
+					Row(
+						modifier = Modifier
+							.wrapContentWidth()
+							.padding(all = 5.dp)
+					) {
+						Text(
+							text = getSecondaryButtonLabel(account.id, bonus, LocalContext.current),
+							style = MaterialTheme.typography.button,
+							modifier = Modifier.padding(end = 5.dp),
+							textAlign = TextAlign.Start,
+							fontSize = 14.sp
+						)
+
+						if(isSimSupportedAccount(account.id)) {
+							Spacer(modifier = Modifier.width(3.dp))
+
+							Image(
+								painter = painterResource(id = R.drawable.ic_bonus),
+								contentDescription = null,
+								modifier = Modifier.size(18.dp)
+							)
+						}
+					}
+				}
+			}
+		}
+	}
+}
+
+private fun getSecondaryButtonLabel(accountId : Int, bonus: Int, context: Context) : String {
+	var label = context.getString(R.string.email_support)
+	if(isSimSupportedAccount(accountId)) {
+		label = context.getString( R.string.nav_airtime)
+		if (bonus > 0) {
+			val bonusPercent = bonus.toString().plus("%")
+			label = context.getString(R.string.buy_airitme_with_discount, bonusPercent)
+		}
+	}
+	return label
+}
+
+@Composable
+private fun SimItemTopRow(simIndex: Int, account: Account, balanceTapListener: BalanceTapListener?) {
+	val size34 = dimensionResource(id = R.dimen.margin_34)
+
+	Row {
+		AsyncImage(
+			model = ImageRequest.Builder(LocalContext.current).data(account.logoUrl)
+				.crossfade(true).diskCachePolicy(CachePolicy.ENABLED).build(),
+			contentDescription = "",
+			placeholder = painterResource(id = R.drawable.img_placeholder),
+			error = painterResource(id = R.drawable.img_placeholder),
+			modifier = Modifier
+				.size(size34)
+				.clip(CircleShape)
+				.align(Alignment.CenterVertically),
+			contentScale = ContentScale.Crop
+		)
+
+		Column(
+			modifier = Modifier
+				.weight(1f)
+				.padding(horizontal = 13.dp)
+		) {
+			Text(text = account.name, style = MaterialTheme.typography.body1)
+			Text(
+				text = stringResource(id = R.string.sim_index, simIndex),
+				color = TextGrey,
+				style = MaterialTheme.typography.body2
+			)
+		}
+
+		Button(
+			onClick = { balanceTapListener?.onTapBalanceRefresh(account) },
+			modifier = Modifier
+				.weight(1f)
+				.shadow(elevation = 0.dp),
+			shape = MaterialTheme.shapes.medium,
+			colors = ButtonDefaults.buttonColors(
+				backgroundColor = BrightBlue,
+				contentColor = ColorPrimary
+			)
+		) {
+			Text(
+				text = stringResource(id = if(isSimSupportedAccount(account.id)) R.string.check_balance_capitalized else R.string.unsupported),
+				style = MaterialTheme.typography.body2,
+				modifier = Modifier.fillMaxWidth(),
+				textAlign = TextAlign.Center
+			)
+		}
+	}
+}
+
+@Composable
+@Preview
+internal fun SimItemsPreview() {
+	StaxTheme {
+		Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colors.background) {
+			Scaffold(topBar = {
+				TopBar(title = R.string.nav_home, isInternetConnected = false, {})
+			}, content = { innerPadding ->
+				val paddingModifier = Modifier.padding(innerPadding)
+				Column {
+
+				}
+			})
+		}
+	}
+}
