+++ app/src/main/res/values/strings.xml
@@ -191,6 +191,8 @@
     <string name="unsupported_simcard_info"><![CDATA[ We detected a <font color="#39CBFC"> %1$s </font> SIM card in slot ( %2$d ), but do not yet support the network.<br/><br/>Please reach out to our support team, so we can integrate with it.]]></string>
     <string name="sim_card_support_request_emailSubject">Please, support my SIM card on Stax</string>
     <string name="sim_card_support_request_emailBody">Autogenerated network provider information (Please do not edit):\n\n OS Reported HNI: %1$s.\n Network name: %2$s,\n Network operator: %3$s,\n Network country: %4$s </string>
+    <string name="unsupported">Unsupported</string>
+    <string name="unknown">Unknown</string>
 
     <string name="search">Search</string>
     <string name="no_accounts_found_desc"><![CDATA[ It looks like "<b><font color="#39CBFC"> %1$s </font> </b>" has not been listed on Stax yet or its an invalid entry.<br/>Try searching for something else.]]></string>
